+++ pkg/combustion/registry.go
@@ -9,6 +9,8 @@ import (
 	"slices"
 	"strings"
 
+	"github.com/schollz/progressbar/v3"
+
 	"gopkg.in/yaml.v3"
 
 	"github.com/suse-edge/edge-image-builder/pkg/fileio"
@@ -360,7 +362,8 @@ func registryArtefactsPath(ctx *image.Context) string {
 }
 
 func populateRegistry(ctx *image.Context, images []string) error {
-	log.Audit("Populating Embedded Artifact Registry...")
+	bar := progressbar.Default(int64(len(images)), "Populating Embedded Artifact Registry...")
+
 	for _, i := range images {
 		if err := addImageToHauler(ctx, i); err != nil {
 			return fmt.Errorf("adding image to hauler: %w", err)
@@ -372,6 +375,10 @@ func populateRegistry(ctx *image.Context, images []string) error {
 		imageTarDest := filepath.Join(registryArtefactsPath(ctx), convertedImageName)
 		if err := generateRegistryTar(ctx, imageTarDest); err != nil {
 			return fmt.Errorf("generating hauler store tar: %w", err)
+		}
+
+		if err := bar.Add(1); err != nil {
+			zap.S().Errorf("Error incrementing the progress bar: %s", err)
 		}
 	}